requirements.txt
pyTelegramBotAPI
python-dotenv
pytz
wikipedia
Pillow
Flask
waitress
requests